Implements kludge for var overrides
The freedomofpress.elk role depends on freedomofpress.generate-ssl-cert.
As such, the vars fed to the ssl role come from elk/meta/main.yml, and
are quite difficult to override. Will need to reconsider the dependency
structure, but for now, adding the vars to the elk role's vars/
directory allows them to take precendence. Further reading:

http://docs.ansible.com/ansible/playbooks_variables.html#variable-scopes